refactor config dir
