+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
  "flake8>=7.2.0",
  "mcp[cli]>=1.8.1",
  "nbformat>=5.10.4",
+ "pydantic-settings>=2.9.1",
  "singlestoredb>=1.12.0",
 ]
