+++ src/config/__init__.py
@@ -1,0 +1,4 @@
+from .app_config import app_config
+from .auth_config import AuthMethod
+
+__all__ = ["app_config", "AuthMethod"]
