+++ src/scripts/init.py
@@ -4,7 +4,7 @@ import sys
 from pathlib import Path
 from typing import Optional, Literal
 from src.auth.auth import get_authentication_token
-from src.config.app_config import AuthMethod, app_config
+from src.config import app_config, AuthMethod
 
 # Supported client types
 ClientType = Literal["claude", "cursor", "windsurf", "copilot"]
